♻️ refactor: extract date format function to form utils
- Moved date formatting logic from DateInput to form utils
- Unified date display format to 'yyyy.mm.dd' across the form
