+++ src/shared/utils/date.js
@@ -7,7 +7,7 @@ export function getFormatDateToString(date) {
   const month = String(date.getMonth() + 1).padStart(2, '0');
   const day = String(date.getDate()).padStart(2, '0');
 
-  return `${year}-${month}-${day}`;
+  return `${year}.${month}.${day}`;
 }
 
 // example : "2025-04-22T13:00:00Z"
